[cineExpress] Separating routes and controllers
